+++ index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="./styles.css">
+  <link rel="stylesheet" href="./src/styles/styles.css">
   <title>YardSale: Tienda Online</title>
 </head>
 <body>
@@ -173,6 +173,6 @@
 
     </div>
   </section>
-  <script src="./main.js"></script>
+  <script src="./src/main.js" type="module"></script>
 </body>
 </html>
