+++ src/components/isCarClose.js
@@ -1,0 +1,11 @@
+const productCart = document.querySelector('.product-cart')
+
+
+const isCarClose = () => {
+  const isproductCartClose = productCart.classList.contains('inactive')
+  if(!isproductCartClose) {
+    productCart.classList.add('inactive')
+  }
+}
+
+export {isCarClose}
