+++ src/main.js
@@ -1,0 +1,15 @@
+import { renderProducts } from "./components/renderProducts.js"
+import {navEmailToggle} from "./components/navEmail.js"
+import {iconCart} from "./components/menuIconCart.js"
+import {burgerMenu} from "./components/burgerMenuIcon.js"
+import {productDetailClose} from "./components/productDetailClose.js"
+
+renderProducts()
+navEmailToggle()
+iconCart()
+burgerMenu()
+productDetailClose()
+
+
+
+
